overhauled the code to correct the missing/corrupt GPS data

+++ build_locations.py
@@ -12,7 +12,7 @@ import pandas as pd
 from ingest import gencsv, fix, rowpred, baypred, rename
 
 
-def main(image_dir, file_path, vineyard, block, date, num_rows, labeled_data):
+def main(image_dir, file_path, vineyard, block, date, num_rows, labeled_data, dry_run):
 	"""
 	Ingests the images indicated by the passed args object.
 	"""
@@ -26,20 +26,28 @@ def main(image_dir, file_path, vineyard, block, date, num_rows, labeled_data):
 		print("Creating the CSV")
 		data = gencsv.create_csv(vineyard, block, date, image_dir)
 
+		#TODO remove
+		data.to_csv("pre_fix.csv")
+
 		print("Fill in the missing GPS data")
 		fix.predict(data)
+
+		#TODO remove
+		data.to_csv("post_fix.csv")
 		
-		print("Predicting the row")
-		rowpred.predict(data, num_rows)
+		if not dry_run:
+			
+			print("Predicting the row")
+			rowpred.predict(data, num_rows)
 
-		print("Predicting the bay")
-		baypred.predict(data, labeled)
+			print("Predicting the bay")
+			baypred.predict(data, labeled)
 
-		print("Renaming")
-		rename.rename(data, date)
+			print("Renaming")
+			rename.rename(data, date)
 
-		# write out the CSV file
-		data.to_csv(file_path)
+			# write out the CSV file
+			data.to_csv(file_path)
 		
 	else:
 		print("Chosen path does not contain files")
@@ -62,14 +70,15 @@ if __name__ == "__main__":
 	ap = ArgumentParser()
 
 	ap.add_argument("-i", "--image_dir", required=True, help="The directory of images")
+	ap.add_argument("-d", "--date", required=True, help="date to ingest (yyyy-mm-dd)")
+	ap.add_argument("-l", "--labeled_data", required=True, help="The locations file labeled row and bay")
+	
 	ap.add_argument("-f", "--file_path", default="locations.csv", help="The path of the CSV file to create")
 	ap.add_argument("-v", "--vineyard", default="crawford-beck", help="vineyard name")
 	ap.add_argument("-b", "--block", default=9, type=int, help="block number")
-	ap.add_argument("-d", "--date", required=True, help="date to ingest (yyyy-mm-dd)")
-	ap.add_argument("-l", "--labeled_data", required=True, help="The locations file labeled row and bay")
 
 	ap.add_argument("-r", "--num_rows", default=21, help="number of rows in block")
-	ap.add_argument("-raw_dir", help="directory of unprocessed images")
+	ap.add_argument("-s", "--dry_run", action="store_true", help="Do not save the results")
 
 	args = ap.parse_args()
